Phy: Fix packet header scan
When looking for the start of a packet, we look for 0x53 in the byte
stream and then look for the rest of the bytes as a valid OSDP packet.
When the packet does not make sense, we just drop what we collected from
the channel thus far (one header width) and continue waiting for another
header.

This is problematic if the 0x53 happens to be in the last few bytes of a
packet which means we would have read the beginning on the next packet
and then dropped that too causing a potentially valid packet to be
dropped. To fix this, let's scan the bytes in cache for another 0x53 and
see if that yields a valid packet.

Fixes: #222

+++ src/osdp_common.h
@@ -407,8 +407,8 @@ struct osdp_pd {
 	/* Raw bytes received from the serial line for this PD */
 	struct osdp_rb rx_rb;
 	uint8_t packet_buf[OSDP_PACKET_BUF_SIZE];
-	int packet_len;
-	int packet_buf_len;
+	unsigned long packet_len;
+	unsigned long packet_buf_len;
 	uint32_t packet_scan_skip;
 
 	int cmd_id;            /* Currently processing command ID */
